+++ test-helpers/fail/array-callback-return.js
@@ -1,0 +1,4 @@
+// In v3, this is added from standard and no longer explicitly needed to be set
+export const badMap = [1, 2, 3].map(n => {
+  console.log(n);
+});