+++ test-helpers/fail/no-var.js
@@ -1,0 +1,2 @@
+// In v3, this is added from standard and no longer explicitly needed to be set
+var a = 42;